Calculate the LO data for all the stuff and Bestandsaufnahme.
--------------------------------------------------
Overview:
--------------------------------------------------
The goal was to use use the nCTEQ15 parameters (for as, quark-masses and Qin) for the calculations to conform with them.
Now, as done for HO before, I calculated all the stuff and plots etc for LO.
I put the LO and HO stuff all in the Bestandsaufnahme_2022_09_2022.

--------------------------------------------------
Changes per file:
--------------------------------------------------
myLHAPDFModified/share/LHAPDF/GRVCustomSetLO/GRVCustomSetLO.info:
    - changed the data to the nCTEQ15 parameters/values

testModifiedApfel/EvolutionFlavors.cc:
    - added LO stuff
    - only small change

results/Bestandsaufnahme_2022_09_15/:
    - added Bestandsaufnahme to collect the LO and HO data

--------------------------------------------------
delted because not used anymore:
--------------------------------------------------
    - testModifiedApfel/EvolutionFlavorsInitialPDFs.cc

--------------------------------------------------
usual changes:
--------------------------------------------------
    - minimizationMinuit/configMinuit.h
    - plottingPython/data_EvolutionFlavors.txt
    - plottingPython/data_InitialPDFs.txt