removed assert in betafunction because it isn't necessary
checking, that the parameters are bigger or equal to zero is not
necessary for the gammafunctions, as long as they aren't negative
integers. Therefore, I will still keep the lower limit of -1 for some of
the parameters, but they actually can be smaller than -1 for the
ErrorPDFs!

also added some Debug stuff